+++ src/skiptoContent.js
@@ -17,7 +17,7 @@ export default class SkipToContent extends HTMLElement {
     super();
     this.attachShadow({ mode: 'open' });
     this.skipToId = 'id-skip-to';
-    this.version = "5.6.0";
+    this.version = "5.6.1";
     this.buttonSkipTo = false;
     this.initialized = false;
 
@@ -87,12 +87,34 @@ export default class SkipToContent extends HTMLElement {
   }
 
   static get observedAttributes() {
-    return ["data-skipto"];
+    return [
+      "data-skipto",
+      "setfocus"
+      ];
   }
 
   attributeChangedCallback(name, oldValue, newValue) {
     if (name === 'data-skipto') {
       this.config = this.setupConfigFromDataAttribute(this.config, newValue);
+    }
+
+    if (name === 'setfocus') {
+        switch(newValue) {
+          case 'button':
+            this.buttonSkipTo.closePopup();
+            this.buttonSkipTo.buttonNode.focus();
+            break;
+
+          case 'menu':
+            this.buttonSkipTo.openPopup();
+            this.buttonSkipTo.setFocusToFirstMenuitem();
+            break;
+
+          case 'none':
+            this.buttonSkipTo.closePopup();
+            document.body.focus();
+            break;
+        }
     }
   }
 
@@ -105,7 +127,7 @@ export default class SkipToContent extends HTMLElement {
    * @param  {object} globalConfig - Reference to configuration object
    *                                 can be undefined
    */
-  init(globalConfig=null) {
+  init(globalConfig=false) {
     if (!this.initialized) {
       this.initialized = true;
       if (globalConfig) {
@@ -121,8 +143,10 @@ export default class SkipToContent extends HTMLElement {
 
       // Add skipto style sheet to document
       renderStyleElement(this.shadowRoot, this.config, this.skipToId);
-      this.buttonSkipTo = new SkiptoMenuButton(this.shadowRoot, this.config, this.skipToId);
+      this.buttonSkipTo = new SkiptoMenuButton(this);
     }
+
+    this.setAttribute('focus', 'none');
   }
 
  /*
@@ -211,6 +235,32 @@ export default class SkipToContent extends HTMLElement {
     if (this.buttonSkipTo) {
       this.buttonSkipTo.updateLabels(config);
     }
+
+    this.setDisplayOption(config['displayOption'], this.skipToId);
+
     return config;
+  }
+
+  /*
+   * @method setDisplayOption
+   *
+   * @desc Set display option for button visibility wehn it does not
+   *       have focus
+   *
+   * @param  {String}  value - String with configuration information
+   * @param  {String}  id    - Id of container element
+   */
+  setDisplayOption(value, id) {
+
+    console.log(`[setDisplayOption]: ${value} ${id}`);
+    const elem = this.shadowRoot.getElementById(id);
+    console.log(`[setDisplayOption][elem]: ${elem}`);
+    if (elem) {
+      elem.classList.remove('popup-border');
+      elem.classList.remove('fixed');
+      elem.classList.remove('popup');
+      elem.classList.remove('static');
+      elem.classList.add(value);
+    }
   }
 }